+++ src/services/workflow-execution.service.ts
@@ -8,21 +8,16 @@ import {
 	ThreadType,
 	type UserWorkflow,
 	type WorkflowRenderedBlock,
-	type WorkflowResponseBlock,
-	type WorkflowTableBlock,
-	type WorkflowTask,
 	type WorkflowTaskResult,
-	type WorkflowTextBlock,
 } from "@/types/memory.js";
 import type { StreamEvent } from "@/types/stream.js";
 import { loggers } from "@/utils/logger.js";
 import type { QueryService } from "./query.service.js";
 import type { ToolCallingService } from "./tool-calling.service.js";
 import type { UserWorkflowService } from "./user-workflow.service.js";
-import {
-	type WorkflowTableRenderResult,
-	WorkflowTableService,
-} from "./workflow-table.service.js";
+import { WorkflowResponseComposer } from "./workflow-response-composer.service.js";
+import { WorkflowTableService } from "./workflow-table.service.js";
+import { WorkflowTaskRunner } from "./workflow-task-runner.service.js";
 import type { WorkflowVariableResolver } from "./workflow-variable-resolver.service.js";
 
 type WorkflowExecutionResult = {
@@ -34,11 +29,9 @@ export class WorkflowExecutionService {
 	private userWorkflowService: UserWorkflowService;
 	private queryService: QueryService;
 	private workflowVariableResolver: WorkflowVariableResolver;
-	private modelModule: ModelModule;
 	private memoryModule: MemoryModule;
-	private a2aModule?: A2AModule;
-	private toolCallingService: ToolCallingService;
-	private workflowTableService: WorkflowTableService;
+	private workflowTaskRunner: WorkflowTaskRunner;
+	private workflowResponseComposer: WorkflowResponseComposer;
 
 	constructor(
 		userWorkflowService: UserWorkflowService,
@@ -52,11 +45,16 @@ export class WorkflowExecutionService {
 		this.userWorkflowService = userWorkflowService;
 		this.queryService = queryService;
 		this.workflowVariableResolver = workflowVariableResolver;
-		this.modelModule = modelModule;
 		this.memoryModule = memoryModule;
-		this.toolCallingService = toolCallingService;
-		this.a2aModule = a2aModule;
-		this.workflowTableService = new WorkflowTableService();
+		this.workflowTaskRunner = new WorkflowTaskRunner(
+			modelModule,
+			toolCallingService,
+			a2aModule,
+		);
+		this.workflowResponseComposer = new WorkflowResponseComposer(
+			modelModule,
+			new WorkflowTableService(),
+		);
 	}
 
 	async executeWorkflow(
@@ -148,7 +146,11 @@ export class WorkflowExecutionService {
 					promptPreview: task.prompt?.slice(0, 200),
 				},
 			);
-			const stream = this.executeTask(task, thread, taskResults);
+			const stream = this.workflowTaskRunner.executeTask(
+				task,
+				thread,
+				taskResults,
+			);
 			let result = await stream.next();
 			while (!result.done) {
 				yield result.value;
@@ -198,7 +200,10 @@ export class WorkflowExecutionService {
 						block.type === "heading" ? undefined : block.sourceTaskIds,
 				},
 			);
-			const stream = this.renderResponseBlock(block, taskResults);
+			const stream = this.workflowResponseComposer.renderResponseBlock(
+				block,
+				taskResults,
+			);
 			let result = await stream.next();
 			while (!result.done) {
 				if (result.value.event === "text_chunk") {
@@ -334,286 +339,5 @@ export class WorkflowExecutionService {
 		await this.memoryModule
 			.getThreadMemory()
 			?.addMessagesToThread(thread.userId, thread.threadId, [message]);
-	}
-
-	private async *executeTask(
-		task: WorkflowTask,
-		thread: ThreadObject,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): AsyncGenerator<StreamEvent, WorkflowTaskResult, unknown> {
-		const startedAt = Date.now();
-		yield {
-			event: "thinking_process",
-			data: {
-				title: `[워크플로우] 작업 실행: ${task.title}`,
-				description: task.agent
-					? `${task.agent.connectorName} 에이전트에 작업을 위임합니다.`
-					: "로컬에서 작업을 실행합니다.",
-				metadata: {
-					phase: "task",
-					taskId: task.taskId,
-					agent: task.agent,
-				},
-			},
-		};
-
-		try {
-			const content = task.agent
-				? yield* this.executeA2ATask(task, thread, taskResults)
-				: yield* this.executeLocalTask(task, thread, taskResults);
-
-			return {
-				taskId: task.taskId,
-				title: task.title,
-				agent: task.agent,
-				status: "completed",
-				content,
-				startedAt,
-				completedAt: Date.now(),
-			};
-		} catch (error) {
-			const message =
-				error instanceof Error ? error.message : JSON.stringify(error);
-			loggers.agent.error(`Workflow task failed: ${task.taskId}`, { error });
-			return {
-				taskId: task.taskId,
-				title: task.title,
-				agent: task.agent,
-				status: "failed",
-				content: "",
-				error: message,
-				startedAt,
-				completedAt: Date.now(),
-			};
-		}
-	}
-
-	private async *executeLocalTask(
-		task: WorkflowTask,
-		thread: ThreadObject,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): AsyncGenerator<StreamEvent, string, unknown> {
-		const model = this.modelModule.getModel();
-		const messages = model.generateMessages({
-			query: this.buildTaskPrompt(task, taskResults),
-			systemPrompt:
-				"You execute one local workflow task. Use MCP tools when useful, use the provided previous task results as context, and return only the task result.",
-		});
-
-		const tools = await this.toolCallingService.getTools({
-			toolPrompt:
-				"현재 워크플로우 작업에 이 MCP 도구가 필요한 이유와 기대하는 결과를 한국어로 구체적으로 작성하세요.",
-			mode: "mcp",
-		});
-		const stream = this.toolCallingService.run({
-			messages,
-			tools,
-			query: task.prompt,
-			thread,
-			toolChoice: "auto",
-		});
-
-		let content = "";
-		let result = await stream.next();
-		while (!result.done) {
-			if (result.value.event === "text_chunk") {
-				content += result.value.data.delta;
-			}
-			yield result.value;
-			result = await stream.next();
-		}
-		return content;
-	}
-
-	private async *executeA2ATask(
-		task: WorkflowTask,
-		thread: ThreadObject,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): AsyncGenerator<StreamEvent, string, unknown> {
-		if (!this.a2aModule || !task.agent) {
-			throw new Error("A2A module is not configured for this workflow task.");
-		}
-
-		const message = this.buildTaskPrompt(task, taskResults);
-		loggers.agent.debug(`Delegating workflow task via A2A: ${task.taskId}`, {
-			taskId: task.taskId,
-			connectorName: task.agent.connectorName,
-			threadId: thread.threadId,
-			messageLength: message.length,
-			messagePreview: message.slice(0, 500),
-		});
-
-		let content = "";
-		const stream = this.a2aModule.sendTask({
-			connectorName: task.agent.connectorName,
-			message,
-			threadId: thread.threadId,
-			metadata: {
-				type: ThreadType.WORKFLOW,
-				workflowId: thread.workflowId,
-				taskId: task.taskId,
-			},
-		});
-
-		let result = await stream.next();
-		while (!result.done) {
-			if (result.value.event === "thinking_process") {
-				yield result.value;
-			} else if (result.value.event === "text_chunk") {
-				content += result.value.data.delta;
-			}
-			result = await stream.next();
-		}
-
-		return content || result.value;
-	}
-
-	private buildTaskPrompt(
-		task: WorkflowTask,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): string {
-		const previousResults = Object.values(taskResults)
-			.map(
-				(result) =>
-					`[${result.taskId}] ${result.title}\nStatus: ${result.status}\nResult:\n${result.content || result.error || ""}`,
-			)
-			.join("\n\n---\n\n");
-
-		return `${previousResults ? `Previous task results:\n${previousResults}\n\n` : ""}Task:
-${task.prompt}`;
-	}
-
-	private async *renderResponseBlock(
-		block: WorkflowResponseBlock,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): AsyncGenerator<StreamEvent, WorkflowRenderedBlock, unknown> {
-		if (block.type === "heading") {
-			const level = block.level ?? 2;
-			const content = `${"#".repeat(level)} ${block.text}\n\n`;
-			yield { event: "text_chunk", data: { delta: content } };
-			return {
-				blockId: block.blockId,
-				type: block.type,
-				content,
-			};
-		}
-
-		if (block.type === "table") {
-			const rendered = yield* this.renderDeterministicTableBlock(
-				block,
-				taskResults,
-			);
-			return {
-				blockId: block.blockId,
-				type: block.type,
-				content: rendered.content,
-				data: rendered.data,
-			};
-		}
-
-		const content = yield* this.renderGeneratedTextBlock(
-			block,
-			taskResults,
-			"Generate this workflow response text from the task results. Return only the response block content.",
-		);
-
-		const finalContent = content.endsWith("\n\n") ? content : `${content}\n\n`;
-		if (finalContent !== content) {
-			yield { event: "text_chunk", data: { delta: "\n\n" } };
-		}
-
-		return {
-			blockId: block.blockId,
-			type: block.type,
-			content: finalContent,
-		};
-	}
-
-	private async *renderDeterministicTableBlock(
-		block: WorkflowTableBlock,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): AsyncGenerator<StreamEvent, WorkflowTableRenderResult, unknown> {
-		const model = this.modelModule.getModel();
-		const modelOptions = this.modelModule.getModelOptions();
-		const sourceResults = this.getSourceTaskResults(block, taskResults);
-		const messages = model.generateMessages({
-			query: this.workflowTableService.buildExtractionPrompt(
-				block,
-				this.serializeTaskResults(sourceResults),
-			),
-			systemPrompt:
-				"Extract only the requested table source values as valid JSON. Return only JSON.",
-		});
-		const response = await model.fetch(messages, modelOptions);
-		const rawContent = response.content || "{}";
-		const rendered = this.workflowTableService.renderTable(block, rawContent);
-		yield { event: "text_chunk", data: { delta: rendered.content } };
-		return rendered;
-	}
-
-	private async *renderGeneratedTextBlock(
-		block: WorkflowTextBlock,
-		taskResults: Record<string, WorkflowTaskResult>,
-		systemPrompt: string,
-	): AsyncGenerator<StreamEvent, string, unknown> {
-		const model = this.modelModule.getModel();
-		const modelOptions = this.modelModule.getModelOptions();
-		const sourceResults = this.getSourceTaskResults(block, taskResults);
-		const messages = model.generateMessages({
-			query: this.buildBlockPrompt(block, sourceResults),
-			systemPrompt,
-		});
-		const stream = await model.fetchStreamWithContextMessage(
-			messages,
-			[],
-			modelOptions,
-		);
-
-		let content = "";
-		for await (const chunk of stream) {
-			if (chunk.delta?.content) {
-				content += chunk.delta.content;
-				yield {
-					event: "text_chunk",
-					data: { delta: chunk.delta.content },
-				};
-			}
-		}
-
-		return content;
-	}
-
-	private getSourceTaskResults(
-		block: Exclude<WorkflowResponseBlock, { type: "heading" }>,
-		taskResults: Record<string, WorkflowTaskResult>,
-	): WorkflowTaskResult[] {
-		if (!block.sourceTaskIds || block.sourceTaskIds.length === 0) {
-			return Object.values(taskResults);
-		}
-
-		return block.sourceTaskIds
-			.map((taskId) => taskResults[taskId])
-			.filter((result): result is WorkflowTaskResult => Boolean(result));
-	}
-
-	private buildBlockPrompt(
-		block: WorkflowTextBlock,
-		taskResults: WorkflowTaskResult[],
-	): string {
-		const resultsText = this.serializeTaskResults(taskResults);
-		return `Task results:
-${resultsText}
-
-Instructions:
-${block.prompt}`;
-	}
-
-	private serializeTaskResults(taskResults: WorkflowTaskResult[]): string {
-		return taskResults
-			.map(
-				(result) =>
-					`[${result.taskId}] ${result.title}\nStatus: ${result.status}\nResult:\n${result.content || result.error || ""}`,
-			)
-			.join("\n\n---\n\n");
 	}
 }
