+++ src/services/workflow-table/shared.ts
@@ -1,0 +1,220 @@
+import type {
+	WorkflowRenderedTableData,
+	WorkflowTableColumnFormat,
+	WorkflowTableColumnFormatKind,
+} from "@/types/memory.js";
+
+export type NumericCellValue = number | null;
+export type RecordCellValue = string | number | null;
+export type RecordTableRow = Record<string, RecordCellValue>;
+export type MatrixTable = Record<string, Record<string, NumericCellValue>>;
+
+export type ParsedMatrixFormula =
+	| {
+			raw: string;
+			target: string;
+			type: "sum";
+			args: string[];
+	  }
+	| {
+			raw: string;
+			target: string;
+			type: "share";
+			args: [string, string];
+	  }
+	| {
+			raw: string;
+			target: string;
+			type: "ratio";
+			args: [string, string];
+	  }
+	| {
+			raw: string;
+			target: string;
+			type: "delta" | "rate" | "growth";
+			args: [string, string];
+	  };
+
+export type ParsedRecordFormula =
+	| {
+			raw: string;
+			type: "binary";
+			target: string;
+			left: string;
+			operator: "+" | "-" | "*" | "/";
+			right: string;
+	  }
+	| {
+			raw: string;
+			type: "total";
+			columns: string[];
+	  };
+
+export type MatrixDefinition = {
+	layout: "matrix";
+	rowHeader: string;
+	rows: string[];
+	columns: string[];
+	columnFormats: Record<string, WorkflowTableColumnFormat>;
+	sourceRows: string[];
+	sourceColumns: string[];
+	computedRowTargets: Set<string>;
+	computedColumnTargets: Set<string>;
+	comparisonColumnTargets: Set<string>;
+	percentRows: Set<string>;
+	percentColumns: Set<string>;
+	formulas: ParsedMatrixFormula[];
+};
+
+export type RecordDefinition = {
+	layout: "records";
+	columns: string[];
+	columnFormats: Record<string, WorkflowTableColumnFormat>;
+	sourceColumns: string[];
+	computedColumns: Set<string>;
+	percentColumns: Set<string>;
+	formulas: ParsedRecordFormula[];
+	totalFormula?: Extract<ParsedRecordFormula, { type: "total" }>;
+};
+
+export type WorkflowTableDefinition = MatrixDefinition | RecordDefinition;
+
+export type WorkflowTableRenderResult = {
+	content: string;
+	data: WorkflowRenderedTableData;
+};
+
+export type ResolvedColumnFormat = {
+	kind: WorkflowTableColumnFormatKind;
+	grouping: boolean;
+	decimals: number;
+	prefix: string;
+	suffix: string;
+	nullDisplay: string;
+};
+
+export type MatrixCellRef = `${string}::${string}`;
+
+export function isFiniteNumber(value: unknown): value is number {
+	return typeof value === "number" && Number.isFinite(value);
+}
+
+function normalizeNumberishString(value: string): string {
+	const trimmed = value.trim();
+	const negativeParentheses =
+		trimmed.startsWith("(") && trimmed.endsWith(")")
+			? `-${trimmed.slice(1, -1)}`
+			: trimmed;
+	return negativeParentheses
+		.replace(/[%,$₩원]/g, "")
+		.replace(/,/g, "")
+		.trim();
+}
+
+function normalizeTextIdentifierString(value: string): string {
+	const trimmed = value.trim();
+	if (/^\d[\d,]*$/.test(trimmed)) {
+		return trimmed.replace(/,/g, "");
+	}
+	return trimmed;
+}
+
+export function parseNumberish(value: unknown): NumericCellValue {
+	if (value === null || value === undefined || value === "") {
+		return null;
+	}
+
+	if (isFiniteNumber(value)) {
+		return value;
+	}
+
+	if (typeof value !== "string") {
+		return null;
+	}
+
+	const trimmed = value.trim();
+	if (
+		trimmed.length === 0 ||
+		trimmed === "-" ||
+		trimmed.toLowerCase() === "null" ||
+		trimmed.toLowerCase() === "n/a"
+	) {
+		return null;
+	}
+
+	const normalized = normalizeNumberishString(trimmed);
+	if (!/^-?\d+(?:\.\d+)?$/.test(normalized)) {
+		return null;
+	}
+
+	const parsed = Number(normalized);
+	return Number.isFinite(parsed) ? parsed : null;
+}
+
+export function parseRecordCell(
+	value: unknown,
+	format?: WorkflowTableColumnFormat,
+): RecordCellValue {
+	if (value === null || value === undefined || value === "") {
+		return null;
+	}
+
+	if (format?.kind === "text") {
+		return typeof value === "string"
+			? normalizeTextIdentifierString(value)
+			: String(value);
+	}
+
+	if (isFiniteNumber(value)) {
+		return value;
+	}
+
+	if (typeof value !== "string") {
+		return null;
+	}
+
+	const trimmed = value.trim();
+	if (
+		trimmed.length === 0 ||
+		trimmed === "-" ||
+		trimmed.toLowerCase() === "null" ||
+		trimmed.toLowerCase() === "n/a"
+	) {
+		return null;
+	}
+
+	const parsed = parseNumberish(trimmed);
+	return parsed === null ? trimmed : parsed;
+}
+
+export function formatNumber(
+	value: number,
+	options: { grouping: boolean; decimals: number },
+): string {
+	return value.toLocaleString("en-US", {
+		useGrouping: options.grouping,
+		minimumFractionDigits: 0,
+		maximumFractionDigits: options.decimals,
+	});
+}
+
+export function safeDivide(
+	numerator: NumericCellValue,
+	denominator: NumericCellValue,
+	warnings: string[],
+	context: string,
+	multiplier = 1,
+): NumericCellValue {
+	if (numerator === null || denominator === null) {
+		return null;
+	}
+	if (denominator === 0) {
+		warnings.push(`Skipped ${context} because the denominator is 0.`);
+		return null;
+	}
+	return (numerator / denominator) * multiplier;
+}
+
+export function looksPercentLike(value: string): boolean {
+	return /%|Pct|Growth/i.test(value);
+}
